inc/xrt: Add facility for "nullary" device-defined utility methods.
"Nullary" because the user supplies no parameters, but
the method receives the xdev pointer as well as a device-defined
void pointer, which can be used for any purpose. This means
these could even be extended over IPC, with the pointer value
serving as a method number to a single trampoline method.