u/sink: Make sink_queue drop the oldest frame when full
Instead of discarding the incoming frame when the queue is
at capacity, drop the oldest frame possible. I can't
think of a situation where it would be better to work on
stale data in preference to fresh data. If there is, we
can easily add a flag to choose behaviour.