d/remote: Adapt r_hub to Mingw64
Targets using r_hub have to be linked with the winsock2 libi, on Windows.
The "#pragma comment(lib, 'ws2_32.lib')" directive used in r_hub.c
is a VC++-ism. When building with Mingw64 (gcc or clang), a standard link
directive is to be used.

Also fix a forgotten replacement of a call to 'close()' by a call to
'socket_close()'.